Make LLM utils
