+++ py/core/base/__init__.py
@@ -127,6 +127,4 @@ __all__ = [
     "yield_sse_event",
     "dump_collector",
     "dump_obj",
-    "resize_base64_image",
-    "estimate_image_tokens",
 ]
