+++ py/core/base/utils/__init__.py
@@ -7,7 +7,6 @@ from shared.utils import (
     deep_update,
     dump_collector,
     dump_obj,
-    estimate_image_tokens,
     format_search_results_for_llm,
     generate_default_prompt_id,
     generate_default_user_collection_id,
@@ -17,7 +16,6 @@ from shared.utils import (
     generate_id,
     generate_user_id,
     increment_version,
-    resize_base64_image,
     validate_uuid,
     yield_sse_event,
 )
@@ -37,7 +35,6 @@ __all__ = [
     "TextSplitter",
     "validate_uuid",
     "deep_update",
-    "resize_base64_image",
     "estimate_image_tokens",
     "_decorate_vector_type",
     "_get_vector_column_str",
