+++ py/core/providers/llm/anthropic.py
@@ -21,8 +21,8 @@ from anthropic.types import (
     RawMessageStartEvent,
     ToolUseBlock,
 )
+from utils import resize_base64_image
 
-from core.base import resize_base64_image
 from core.base.abstractions import GenerationConfig, LLMChatCompletion
 from core.base.providers.llm import CompletionConfig, CompletionProvider
