+++ py/core/providers/llm/openai.py
@@ -3,8 +3,8 @@ import os
 from typing import Any
 
 from openai import AsyncAzureOpenAI, AsyncOpenAI, OpenAI
+from utils import resize_base64_image
 
-from core.base import resize_base64_image
 from core.base.abstractions import GenerationConfig
 from core.base.providers.llm import CompletionConfig, CompletionProvider
